# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type IocType = typeof IocType[keyof typeof IocType];


export const IocType = {
  ipv4: 'ipv4',
  ipv6: 'ipv6',
  domain: 'domain',
  url: 'url',
  email: 'email',
  md5: 'md5',
  sha1: 'sha1',
  sha256: 'sha256',
  cve: 'cve',
} as const;

/**
 * A single indicator of compromise extracted or referenced.
 */
export interface Ioc {
  type: IocType;
  value: string;
}

export interface MitreMapping {
  tacticId: string;
  tacticName: string;
  techniqueId: string;
  techniqueName: string;
}

export interface LogAnalysisInput {
  /**
     * Raw log text to analyze (syslog, firewall, EDR, auth logs, etc.)
     * @minLength 1
     */
  logText: string;
}

export type LogAnalysisResultSeverity = typeof LogAnalysisResultSeverity[keyof typeof LogAnalysisResultSeverity];


export const LogAnalysisResultSeverity = {
  informational: 'informational',
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface LogAnalysisResult {
  id: string;
  severity: LogAnalysisResultSeverity;
  summary: string;
  indicators: Ioc[];
  mitreMappings: MitreMapping[];
  investigationSteps: string[];
  containmentRecommendations: string[];
  createdAt: string;
}

export interface IncidentReportInput {
  /** @minLength 1 */
  incidentTitle: string;
  /**
     * Free-form description of the incident, affected systems, timeline notes, and any logs/IOCs already gathered.
     * @minLength 1
     */
  details: string;
  /**
     * Optional id of a prior LogAnalysisResult saved in history to use as context.
     * @nullable
     */
  priorAnalysisId?: string | null;
}

export interface TimelineEvent {
  timestamp: string;
  description: string;
}

export interface IncidentReport {
  id: string;
  incidentTitle: string;
  executiveSummary: string;
  timeline: TimelineEvent[];
  indicators: Ioc[];
  affectedAssets: string[];
  rootCause: string;
  impact: string;
  containment: string;
  eradication: string;
  recovery: string;
  lessonsLearned: string;
  createdAt: string;
}

export type HistoryEntryType = typeof HistoryEntryType[keyof typeof HistoryEntryType];


export const HistoryEntryType = {
  log_analysis: 'log_analysis',
  incident_report: 'incident_report',
} as const;

export interface HistoryEntry {
  id: string;
  type: HistoryEntryType;
  title: string;
  input: string;
  /** The full LogAnalysisResult or IncidentReport payload, depending on type. */
  result: unknown;
  createdAt: string;
}