# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntryType } from './historyEntryType';

export interface HistoryEntry {
  id: string;
  type: HistoryEntryType;
  title: string;
  input: string;
  /** The full LogAnalysisResult or IncidentReport payload, depending on type. */
  result: unknown;
  createdAt: Date;
}