# lib/api-zod/src/generated/types/incidentReport.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ioc } from './ioc';
import type { TimelineEvent } from './timelineEvent';

export interface IncidentReport {
  id: string;
  incidentTitle: string;
  executiveSummary: string;
  timeline: TimelineEvent[];
  indicators: Ioc[];
  affectedAssets: string[];
  rootCause: string;
  impact: string;
  containment: string;
  eradication: string;
  recovery: string;
  lessonsLearned: string;
  createdAt: Date;
}