# lib/api-zod/src/generated/types/incidentReportInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface IncidentReportInput {
  /** @minLength 1 */
  incidentTitle: string;
  /**
     * Free-form description of the incident, affected systems, timeline notes, and any logs/IOCs already gathered.
     * @minLength 1
     */
  details: string;
  /**
     * Optional id of a prior LogAnalysisResult saved in history to use as context.
     * @nullable
     */
  priorAnalysisId?: string | null;
}