# lib/api-zod/src/generated/types/logAnalysisInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LogAnalysisInput {
  /**
     * Raw log text to analyze (syslog, firewall, EDR, auth logs, etc.)
     * @minLength 1
     */
  logText: string;
}