# lib/api-zod/src/generated/types/logAnalysisResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ioc } from './ioc';
import type { LogAnalysisResultSeverity } from './logAnalysisResultSeverity';
import type { MitreMapping } from './mitreMapping';

export interface LogAnalysisResult {
  id: string;
  severity: LogAnalysisResultSeverity;
  summary: string;
  indicators: Ioc[];
  mitreMappings: MitreMapping[];
  investigationSteps: string[];
  containmentRecommendations: string[];
  createdAt: Date;
}