# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  timestamp: string;
  description: string;
}